Decompose execution_test.py
Unit tests moved to
- dispatcher_test.py
- runner_test.py
- result_test.py

execution_test.py is reserved for integration tests
