+++ tests/covalent_dispatcher_tests/_core/execution_test.py
@@ -19,35 +19,18 @@
 # Relief from the License may be granted by purchasing a commercial license.
 
 """
-Tests for the core functionality of the dispatcher.
+Integration tests for the dispatcher, runner, and result modules
 """
 
 
-from asyncio import Queue
-from typing import Dict, List
-from unittest.mock import AsyncMock
-
-import cloudpickle as pickle
 import pytest
 
 import covalent as ct
 from covalent._results_manager import Result
 from covalent._shared_files.defaults import sublattice_prefix
 from covalent._workflow.lattice import Lattice
-from covalent_dispatcher._core.dispatcher import (
-    _build_sublattice_graph,
-    _dispatch_sync_sublattice,
-    _get_abstract_task_inputs,
-    _handle_cancelled_node,
-    _handle_completed_node,
-    _handle_failed_node,
-    _initialize_deps_and_queue,
-    _plan_workflow,
-    _post_process,
-    run_workflow,
-)
-from covalent_dispatcher._core.result import _update_node_result, initialize_result_object
-from covalent_dispatcher._core.runner import _gather_deps, _get_task_inputs, _run_task
+from covalent_dispatcher._core.dispatcher import _dispatch_sync_sublattice, run_workflow
+from covalent_dispatcher._core.runner import _run_task
 from covalent_dispatcher._db import update
 from covalent_dispatcher._db.datastore import DataStore
 
@@ -88,478 +71,6 @@ def get_mock_result() -> Result:
     )
 
     return result_object
-
-
-def test_initialize_result_object(mocker, test_db):
-    """Test the `initialize_result_object` function"""
-
-    @ct.electron
-    def task(x):
-        return x
-
-    @ct.lattice
-    def workflow(x):
-        return task(x)
-
-    workflow.build_graph(1)
-    json_lattice = workflow.serialize_to_json()
-    mocker.patch("covalent_dispatcher._db.upsert.workflow_db", return_value=test_db)
-    mocker.patch("covalent_dispatcher._db.write_result_to_db.workflow_db", return_value=test_db)
-    result_object = get_mock_result()
-
-    mock_persist = mocker.patch("covalent_dispatcher._db.update.persist")
-
-    sub_result_object = initialize_result_object(
-        json_lattice=json_lattice, parent_result_object=result_object, parent_electron_id=5
-    )
-
-    mock_persist.assert_called_with(sub_result_object, electron_id=5)
-    assert sub_result_object._root_dispatch_id == result_object.dispatch_id
-
-
-def test_plan_workflow():
-    """Test workflow planning method."""
-
-    @ct.electron
-    def task(x):
-        return x
-
-    @ct.lattice
-    def workflow(x):
-        return task(x)
-
-    workflow.metadata["schedule"] = True
-    received_workflow = Lattice.deserialize_from_json(workflow.serialize_to_json())
-    result_object = Result(received_workflow, "/tmp", "asdf")
-    _plan_workflow(result_object=result_object)
-
-    # Updated transport graph post planning
-    updated_tg = pickle.loads(result_object.lattice.transport_graph.serialize(metadata_only=True))
-
-    assert updated_tg["lattice_metadata"]["schedule"]
-
-
-def test_post_process():
-    """Test post-processing of results."""
-
-    import covalent as ct
-
-    @ct.electron
-    def construct_cu_slab(x):
-        return x
-
-    @ct.electron
-    def compute_system_energy(x):
-        return x
-
-    @ct.electron
-    def construct_n_molecule(x):
-        return x
-
-    @ct.electron
-    def get_relaxed_slab(x):
-        return x
-
-    @ct.lattice
-    def compute_energy():
-        N2 = construct_n_molecule(1)
-        e_N2 = compute_system_energy(N2)
-
-        slab = construct_cu_slab(2)
-        e_slab = compute_system_energy(slab)
-
-        relaxed_slab = get_relaxed_slab(3)
-        e_relaxed_slab = compute_system_energy(relaxed_slab)
-
-        return (N2, e_N2, slab, e_slab, relaxed_slab, e_relaxed_slab)
-
-    compute_energy.build_graph()
-
-    node_outputs = {
-        "construct_n_molecule(0)": 1,
-        ":parameter:1(1)": 1,
-        "compute_system_energy(2)": 1,
-        "construct_cu_slab(3)": 2,
-        ":parameter:2(4)": 2,
-        "compute_system_energy(5)": 2,
-        "get_relaxed_slab(6)": 3,
-        ":parameter:3(7)": 3,
-        "compute_system_energy(8)": 3,
-    }
-
-    encoded_node_outputs = {
-        k: ct.TransportableObject.make_transportable(v) for k, v in node_outputs.items()
-    }
-
-    execution_result = _post_process(compute_energy, encoded_node_outputs)
-
-    assert execution_result == compute_energy()
-
-
-def test_get_task_inputs():
-    """Test _get_task_inputs for both dicts and list parameter types"""
-
-    @ct.electron
-    def list_task(arg: List):
-        return len(arg)
-
-    @ct.electron
-    def dict_task(arg: Dict):
-        return len(arg)
-
-    @ct.electron
-    def multivariable_task(x, y):
-        return x, y
-
-    @ct.lattice
-    def list_workflow(arg):
-        return list_task(arg)
-
-    @ct.lattice
-    def dict_workflow(arg):
-        return dict_task(arg)
-
-    #    1   2
-    #     \   \
-    #      0   3
-    #     / /\/
-    #     4   5
-
-    @ct.electron
-    def identity(x):
-        return x
-
-    @ct.lattice
-    def multivar_workflow(x, y):
-        electron_x = identity(x)
-        electron_y = identity(y)
-        res1 = multivariable_task(electron_x, electron_y)
-        res2 = multivariable_task(electron_y, electron_x)
-        res3 = multivariable_task(electron_y, electron_x)
-        res4 = multivariable_task(electron_x, electron_y)
-        return 1
-
-    # list-type inputs
-
-    list_workflow.build_graph([1, 2, 3])
-    serialized_args = [ct.TransportableObject(i) for i in [1, 2, 3]]
-    tg = list_workflow.transport_graph
-    # Nodes 0=task, 1=:electron_list:, 2=1, 3=2, 4=3
-    tg.set_node_value(2, "output", ct.TransportableObject(1))
-    tg.set_node_value(3, "output", ct.TransportableObject(2))
-    tg.set_node_value(4, "output", ct.TransportableObject(3))
-
-    result_object = Result(lattice=list_workflow, results_dir="/tmp", dispatch_id="asdf")
-    task_inputs = _get_task_inputs(1, tg.get_node_value(1, "name"), result_object)
-
-    expected_inputs = {"args": serialized_args, "kwargs": {}}
-
-    assert task_inputs == expected_inputs
-
-    # dict-type inputs
-
-    dict_workflow.build_graph({"a": 1, "b": 2})
-    serialized_args = {"a": ct.TransportableObject(1), "b": ct.TransportableObject(2)}
-    tg = dict_workflow.transport_graph
-    # Nodes 0=task, 1=:electron_dict:, 2=1, 3=2
-    tg.set_node_value(2, "output", ct.TransportableObject(1))
-    tg.set_node_value(3, "output", ct.TransportableObject(2))
-
-    result_object = Result(lattice=dict_workflow, results_dir="/tmp", dispatch_id="asdf")
-    task_inputs = _get_task_inputs(1, tg.get_node_value(1, "name"), result_object)
-    expected_inputs = {"args": [], "kwargs": serialized_args}
-
-    assert task_inputs == expected_inputs
-
-    # Check arg order
-    multivar_workflow.build_graph(1, 2)
-    received_lattice = Lattice.deserialize_from_json(multivar_workflow.serialize_to_json())
-    result_object = Result(lattice=received_lattice, results_dir="/tmp", dispatch_id="asdf")
-    tg = received_lattice.transport_graph
-
-    assert list(tg._graph.nodes) == [0, 1, 2, 3, 4, 5, 6, 7]
-    tg.set_node_value(0, "output", ct.TransportableObject(1))
-    tg.set_node_value(2, "output", ct.TransportableObject(2))
-
-    task_inputs = _get_task_inputs(4, tg.get_node_value(4, "name"), result_object)
-
-    input_args = [arg.get_deserialized() for arg in task_inputs["args"]]
-    assert input_args == [1, 2]
-
-    task_inputs = _get_task_inputs(5, tg.get_node_value(5, "name"), result_object)
-    input_args = [arg.get_deserialized() for arg in task_inputs["args"]]
-    assert input_args == [2, 1]
-
-    task_inputs = _get_task_inputs(6, tg.get_node_value(6, "name"), result_object)
-    input_args = [arg.get_deserialized() for arg in task_inputs["args"]]
-    assert input_args == [2, 1]
-
-    task_inputs = _get_task_inputs(7, tg.get_node_value(7, "name"), result_object)
-    input_args = [arg.get_deserialized() for arg in task_inputs["args"]]
-    assert input_args == [1, 2]
-
-
-def test_get_abstract_task_inputs():
-    """Test _get_abstract_task_inputs for both dicts and list parameter types"""
-
-    @ct.electron
-    def list_task(arg: List):
-        return len(arg)
-
-    @ct.electron
-    def dict_task(arg: Dict):
-        return len(arg)
-
-    @ct.electron
-    def multivariable_task(x, y):
-        return x, y
-
-    @ct.lattice
-    def list_workflow(arg):
-        return list_task(arg)
-
-    @ct.lattice
-    def dict_workflow(arg):
-        return dict_task(arg)
-
-    #    1   2
-    #     \   \
-    #      0   3
-    #     / /\/
-    #     4   5
-
-    @ct.electron
-    def identity(x):
-        return x
-
-    @ct.lattice
-    def multivar_workflow(x, y):
-        electron_x = identity(x)
-        electron_y = identity(y)
-        res1 = multivariable_task(electron_x, electron_y)
-        res2 = multivariable_task(electron_y, electron_x)
-        res3 = multivariable_task(electron_y, electron_x)
-        res4 = multivariable_task(electron_x, electron_y)
-        return 1
-
-    # list-type inputs
-
-    # Nodes 0=task, 1=:electron_list:, 2=1, 3=2, 4=3
-    list_workflow.build_graph([1, 2, 3])
-    abstract_args = [2, 3, 4]
-    tg = list_workflow.transport_graph
-
-    result_object = Result(lattice=list_workflow, results_dir="/tmp", dispatch_id="asdf")
-    abs_task_inputs = _get_abstract_task_inputs(1, tg.get_node_value(1, "name"), result_object)
-
-    expected_inputs = {"args": abstract_args, "kwargs": {}}
-
-    assert abs_task_inputs == expected_inputs
-
-    # dict-type inputs
-
-    # Nodes 0=task, 1=:electron_dict:, 2=1, 3=2
-    dict_workflow.build_graph({"a": 1, "b": 2})
-    abstract_args = {"a": 2, "b": 3}
-    tg = dict_workflow.transport_graph
-
-    result_object = Result(lattice=dict_workflow, results_dir="/tmp", dispatch_id="asdf")
-    task_inputs = _get_abstract_task_inputs(1, tg.get_node_value(1, "name"), result_object)
-    expected_inputs = {"args": [], "kwargs": abstract_args}
-
-    assert task_inputs == expected_inputs
-
-    # Check arg order
-    multivar_workflow.build_graph(1, 2)
-    received_lattice = Lattice.deserialize_from_json(multivar_workflow.serialize_to_json())
-    result_object = Result(lattice=received_lattice, results_dir="/tmp", dispatch_id="asdf")
-    tg = received_lattice.transport_graph
-
-    assert list(tg._graph.nodes) == [0, 1, 2, 3, 4, 5, 6, 7]
-    tg.set_node_value(0, "output", ct.TransportableObject(1))
-    tg.set_node_value(2, "output", ct.TransportableObject(2))
-
-    task_inputs = _get_abstract_task_inputs(4, tg.get_node_value(4, "name"), result_object)
-    assert task_inputs["args"] == [0, 2]
-
-    task_inputs = _get_abstract_task_inputs(5, tg.get_node_value(5, "name"), result_object)
-    assert task_inputs["args"] == [2, 0]
-
-    task_inputs = _get_abstract_task_inputs(6, tg.get_node_value(6, "name"), result_object)
-    assert task_inputs["args"] == [2, 0]
-
-    task_inputs = _get_abstract_task_inputs(7, tg.get_node_value(7, "name"), result_object)
-    assert task_inputs["args"] == [0, 2]
-
-
-def test_gather_deps():
-    """Test internal _gather_deps for assembling deps into call_before and
-    call_after"""
-
-    def square(x):
-        return x * x
-
-    @ct.electron(deps_bash=ct.DepsBash("ls -l"), call_after=[ct.DepsCall(square, [3])])
-    def task(x):
-        return x
-
-    @ct.lattice
-    def workflow(x):
-        return task(x)
-
-    workflow.build_graph(5)
-
-    received_workflow = Lattice.deserialize_from_json(workflow.serialize_to_json())
-    result_object = Result(received_workflow, "/tmp", "asdf")
-
-    before, after = _gather_deps(result_object, 0)
-    assert len(before) == 1
-    assert len(after) == 1
-
-
-@pytest.mark.asyncio
-async def test_update_failed_node(mocker):
-    """Check that update_node_result correctly invokes _handle_failed_node"""
-
-    status_queue = AsyncMock()
-
-    pending_deps = {}
-
-    result_object = get_mock_result()
-    mock_fail_handler = mocker.patch("covalent_dispatcher._core.dispatcher._handle_failed_node")
-    mock_upsert_lattice = mocker.patch("covalent_dispatcher._db.upsert._lattice_data")
-    mock_update_node = mocker.patch("covalent_dispatcher._db.update._node")
-
-    node_result = {"node_id": 0, "status": Result.FAILED}
-    await _update_node_result(result_object, node_result, pending_deps, status_queue)
-
-    status_queue.put.assert_awaited_with((0, Result.FAILED))
-
-
-@pytest.mark.asyncio
-async def test_update_cancelled_node(mocker):
-    """Check that update_node_result correctly invokes _handle_cancelled_node"""
-
-    status_queue = AsyncMock()
-
-    pending_deps = {}
-
-    result_object = get_mock_result()
-    mock_cancel_handler = mocker.patch(
-        "covalent_dispatcher._core.dispatcher._handle_cancelled_node"
-    )
-    mock_upsert_lattice = mocker.patch("covalent_dispatcher._db.upsert._lattice_data")
-    mock_update_node = mocker.patch("covalent_dispatcher._db.update._node")
-
-    node_result = {"node_id": 0, "status": Result.CANCELLED}
-    await _update_node_result(result_object, node_result, pending_deps, status_queue)
-    status_queue.put.assert_awaited_with((0, Result.CANCELLED))
-
-
-@pytest.mark.asyncio
-async def test_update_completed_node(mocker):
-    """Check that update_node_result correctly invokes _handle_completed_node"""
-
-    status_queue = AsyncMock()
-    pending_deps = {}
-
-    result_object = get_mock_result()
-    mock_completed_handler = mocker.patch(
-        "covalent_dispatcher._core.dispatcher._handle_completed_node"
-    )
-    mock_upsert_lattice = mocker.patch("covalent_dispatcher._db.upsert._lattice_data")
-    mock_update_node = mocker.patch("covalent_dispatcher._db.update._node")
-
-    node_result = {"node_id": 0, "status": Result.COMPLETED}
-    await _update_node_result(result_object, node_result, pending_deps, status_queue)
-    status_queue.put.assert_awaited_with((0, Result.COMPLETED))
-
-
-@pytest.mark.asyncio
-async def test_handle_completed_node(mocker):
-    """Unit test for completed node handler"""
-    status_queue = Queue()
-    pending_deps = {}
-
-    result_object = get_mock_result()
-
-    # tg edges are (1, 0), (0, 2)
-    pending_deps[0] = 1
-    pending_deps[1] = 0
-    pending_deps[2] = 1
-
-    mock_upsert_lattice = mocker.patch("covalent_dispatcher._db.upsert._lattice_data")
-
-    node_result = {"node_id": 1, "status": Result.COMPLETED}
-
-    next_nodes = await _handle_completed_node(result_object, 1, pending_deps)
-    assert next_nodes == [0]
-    assert pending_deps == {0: 0, 1: 0, 2: 1}
-
-
-@pytest.mark.asyncio
-async def test_handle_failed_node(mocker):
-    """Unit test for failed node handler"""
-    status_queue = Queue()
-    pending_deps = {}
-
-    result_object = get_mock_result()
-
-    # tg edges are (1, 0), (0, 2)
-    pending_deps[0] = 1
-    pending_deps[1] = 0
-    pending_deps[2] = 1
-
-    mock_upsert_lattice = mocker.patch("covalent_dispatcher._db.upsert._lattice_data")
-    mock_get_node_name = mocker.patch("covalent._results_manager.result.Result._get_node_name")
-
-    mock_get_node_error = mocker.patch("covalent._results_manager.result.Result._get_node_error")
-
-    node_result = {"node_id": 1, "status": Result.FAILED}
-
-    await _handle_failed_node(result_object, 1)
-
-    assert pending_deps == {0: 1, 1: 0, 2: 1}
-    assert result_object.status == Result.FAILED
-    mock_get_node_name.assert_called_once()
-    mock_get_node_error.assert_called_once()
-
-
-@pytest.mark.asyncio
-async def test_handle_cancelled_node(mocker):
-    """Unit test for cancelled node handler"""
-    status_queue = Queue()
-    pending_deps = {}
-
-    result_object = get_mock_result()
-
-    # tg edges are (1, 0), (0, 2)
-    pending_deps[0] = 1
-    pending_deps[1] = 0
-    pending_deps[2] = 1
-
-    mock_upsert_lattice = mocker.patch("covalent_dispatcher._db.upsert._lattice_data")
-
-    node_result = {"node_id": 1, "status": Result.CANCELLED}
-
-    await _handle_cancelled_node(result_object, 1)
-
-    assert pending_deps == {0: 1, 1: 0, 2: 1}
-    assert result_object.status == Result.CANCELLED
-
-
-@pytest.mark.asyncio
-async def test_initialize_deps_and_queue(mocker):
-    """Test internal function for initializing status_queue and pending_deps"""
-    status_queue = Queue()
-    pending_deps = {}
-
-    result_object = get_mock_result()
-    num_tasks, initial_nodes, pending_deps = await _initialize_deps_and_queue(result_object)
-
-    assert initial_nodes == [1]
-    assert pending_deps == {0: 1, 1: 0, 2: 1}
-    assert num_tasks == len(result_object.lattice.transport_graph._graph.nodes)
 
 
 @pytest.mark.asyncio
@@ -743,21 +254,6 @@ async def test_run_workflow_with_failed_postprocess(test_db, mocker):
     result_object = await run_workflow(result_object)
 
     assert result_object.status == Result.POSTPROCESSING_FAILED
-
-
-def test_build_sublattice_graph():
-    @ct.electron
-    def task(x):
-        return x
-
-    @ct.lattice
-    def workflow(x):
-        return task(x)
-
-    json_lattice = _build_sublattice_graph(workflow, 1)
-    lattice = Lattice.deserialize_from_json(json_lattice)
-
-    assert list(lattice.transport_graph._graph.nodes) == [0, 1]
 
 
 @pytest.mark.asyncio
